\name{util_cleanCredentials}
\alias{util_cleanCredentials}
\title{Clean API Credentials}
\usage{util_cleanCredentials()
}
\description{
Removes existing records of client API credentials from the system.
}

\author{Kostin Andrey <andrey.kostin@portfolioeffect.com>}
\seealso{\code{\link{portfolio_create}} }
\examples{
\dontrun{ 
util_cleanCredentials()
}}
\keyword{PortfolioEffect}
\keyword{util_cleanCredentials}